revno: 3175 message:   (local vn) Excel reporting - Updated by data element service
committer: Hieu <hieu.hispvietnam@gmail.com>
branch nick: dhis2
timestamp: Mon 2011-03-28 13:49:08 +0700

+++ local/vn/dhis-web-excel-reporting/src/main/java/org/hisp/dhis/reportexcel/utils/ExpressionUtils.java
@@ -99,11 +99,10 @@ public class ExpressionUtils
                 {
                     CalculatedDataElement calculatedDataElement = (CalculatedDataElement) dataElement;
 
-                    int factor = 0;
-
+                    double factor = 0;
                     double value = 0.0;
 
-                    Map<String, Integer> factorMap = dataElementService.getOperandFactors( calculatedDataElement );
+                    Map<String, Double> factorMap = dataElementService.getOperandFactors( calculatedDataElement );
 
                     Collection<String> operandIds = dataElementService.getOperandIds( calculatedDataElement );
